requirements.txt
llama-index==0.10.*
python-dotenv==1.*
chromadb
llama-index-vector-stores-chroma